Added Ubuntu support by recompiling with -fPIE
In order for the generated libraries to be linkable in Ubuntu and
runnable in CentOS, they must be compiled as Position-Independent
Executables (PIE). This is done by injecting the -fPIE cflag in when
compiling ImageMagick. The updated object files in this commit were
generated with this flag added.

Additional changes were made to documentation and Vagrantfile
in order to test building in Ubuntu, linking in Ubuntu via Bazel, and
running the final binary in CentOS.

The version number referenced in the README was also corrected to match
the version downloaded/built in download-build-imagemagick.sh.